requirements.txt
# Core agent deps
langchain
langchain-openai
langgraph
pydantic
python-dotenv
sqlalchemy
pymysql
bcrypt
python-jose[cryptography]

# Product retrieval (teammate's ProductRetrievalService)
faiss-cpu
transformers
torch
torchvision
pandas
numpy
huggingface_hub

# Gradio UI (existing app.py)
gradio

# FastAPI backend (api_clip.py)
fastapi
uvicorn[standard]
python-multipart
